requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.7
numpy==1.26.4
scikit-learn==1.3.2
google-generativeai==0.12.0
python-dotenv==1.0.1
gunicorn==21.2.0

# Flask==3.0.3
# Werkzeug>=3.0.0
# itsdangerous>=2.1.2
# Jinja2>=3.1.2
# click>=8.1.7
# numpy>=1.26.0
# scikit-learn>=1.3.0
# google-generativeai>=0.12.0
# python-dotenv>=1.0.0